% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent_fxns.R
\name{assign_roles}
\alias{assign_roles}
\title{Assign roles to agents}
\usage{
assign_roles(input_df, module_leads, prop_overhead_leads)
}
\arguments{
\item{input_df}{data frame containing agent data}

\item{module_leads}{number of leads per module}

\item{overhead_leads}{number of leads per overhead module}
}
\description{
Assign roles to agents
}
