% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean_up.R
\name{clean_mods}
\alias{clean_mods}
\title{Clean up module data}
\usage{
clean_mods(mod_id_df)
}
\arguments{
\item{mod_id_df}{Module id assignment data frame}
}
\value{
Module ID data frame where all overhead assignments are changed to "O-100"
}
\description{
The high number of unique overhead module assignments was affecting spread
dynamics. This function solves the issue by grouping all overhead into a
single module designated "O-100"
}
