% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean_up.R
\name{count_daily_inf}
\alias{count_daily_inf}
\title{Count daily new infections}
\usage{
count_daily_inf(new_df, agent_df)
}
\arguments{
\item{new_df}{Data frame from current/new day}

\item{agent_df}{Data frame from previous day}
}
\value{
A summary data frame that counts new infections by time, inc_id,
quarantine status, and vaccinated status
}
\description{
Count daily new infections
}
