% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_fxns.R
\name{expose_leads}
\alias{expose_leads}
\title{Find exposed leaders by unique fire (inc_id)}
\usage{
expose_leads(input_df, BI, BA, exp_thres = 0, delta_t = 1)
}
\arguments{
\item{input_df}{Input data frame}

\item{BI}{Beta for Symptomatic}

\item{BA}{Beta for Asymptomatic}

\item{exp_thres}{Exposure threshold}

\item{delta_t}{Time step}
}
\value{
Vector of resource ID's that have been newly exposed
}
\description{
Find exposed leaders by unique fire (inc_id)
}
