% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_fxns.R
\name{expose_off_fire}
\alias{expose_off_fire}
\title{Find exposed agents that are off fire}
\usage{
expose_off_fire(input_df, eir)
}
\arguments{
\item{input_df}{complete agent data frame (all incidents and modules)}

\item{eir}{entry infection rate, or off fire exposure rate}
}
\description{
Find exposed agents that are off fire
}
