% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_fxns.R
\name{find_exposed_in_leads}
\alias{find_exposed_in_leads}
\title{Find exposed leads within a single incident}
\usage{
find_exposed_in_leads(inc_df, BA, BI, exp_thres = 0, delta_t = 1)
}
\arguments{
\item{inc_df}{Data frame containing agent information for a single incident}

\item{BA}{Beta for Asymptomatic}

\item{BI}{Beta for Symptomatic}

\item{exp_thres}{Exposure threshold}

\item{delta_t}{Time step}
}
\value{
Vector of resource IDs that have been exposed
}
\description{
Find exposed leads within a single incident
}
