% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_fxns.R
\name{find_exposed_in_mod}
\alias{find_exposed_in_mod}
\title{Find exposed agents within a single module}
\usage{
find_exposed_in_mod(mod_df, BA, BI, multiplier, exp_thres = 0, delta_t = 1)
}
\arguments{
\item{mod_df}{Data frame containing agent information for a single module}

\item{BA}{Beta for Asymptomatic}

\item{BI}{Beta for Symptomatic}

\item{multiplier}{scalar for increasing R0 within modules}

\item{exp_thres}{Exposure threshold}

\item{delta_t}{Time step}
}
\value{
Vector of resource ID's that have been newly exposed
}
\description{
Find exposed agents within a single module
}
