% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean_up.R
\name{find_first_mob}
\alias{find_first_mob}
\title{Find the first mobilization for an agent/res_id}
\usage{
find_first_mob(res_data)
}
\arguments{
\item{res_data}{Vector of all incident assignments for a single res_id}
}
\value{
First mobilization found in the supplied assignment vector
}
\description{
Find the first mobilization for an agent/res_id
}
