% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_role_type.R
\name{find_role_type}
\alias{find_role_type}
\title{Assign each firefighter agent a role "type"}
\usage{
find_role_type(mod_data)
}
\arguments{
\item{mod_data}{module assignment data}
}
\value{
character vector of agent role/module "types" (Overhead, OverheadCrew,
OverheadEngine, OverheadCrewEngine, Crew, CrewEngine, Engine)
}
\description{
Inspired by a fizzbuzz, this function finds which module types each agent has
been on and prints them out
}
\examples{
\dontrun{
find_role_type(mod_id_2017)
}
}
