% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/increment.R
\name{\%+=\%}
\alias{\%+=\%}
\title{I'm just petty and wanted an increment operator...}
\usage{
e1 \%+=\% e2
}
\arguments{
\item{e1}{...}

\item{e2}{...}
}
\description{
I'm just petty and wanted an increment operator...
}
