% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean_up.R
\name{inc_id_to_gacc}
\alias{inc_id_to_gacc}
\title{Translate inc_id into inc_gacc}
\usage{
inc_id_to_gacc(inc_id, inc_info)
}
\arguments{
\item{inc_id}{numeric incident ID}

\item{inc_info}{data frame containing seasonal information on incidents}
}
\description{
Translate inc_id into inc_gacc
}
