% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent_fxns.R
\name{inc_subsetter}
\alias{inc_subsetter}
\title{Subset agent dataset by incident}
\usage{
inc_subsetter(agent_df)
}
\arguments{
\item{agent_df}{data frame containing agent data}
}
\description{
Subset agent dataset by incident
}
