% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent_fxns.R
\name{mk_agents}
\alias{mk_agents}
\title{Initializes the agent data frame}
\usage{
mk_agents(inc_req, mod_req, day = 1)
}
\arguments{
\item{inc_req}{data frame containing incident assignments}

\item{mod_req}{data frame containing module assignments}

\item{day}{day of the season to use as the initial data frame, default is 1}
}
\description{
Initializes the agent data frame
}
