% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent_fxns.R
\name{mod_subsetter}
\alias{mod_subsetter}
\title{Subset agent data frame by module}
\usage{
mod_subsetter(agent_df)
}
\arguments{
\item{agent_df}{data frame containing agent data}
}
\value{
A list object, each entry is a tibble containing data for an individual
module.
}
\description{
Subset agent data frame by module
}
