% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modular_quarantine.R
\name{modular_quarantine}
\alias{modular_quarantine}
\title{Module based quarantine operations}
\usage{
modular_quarantine(input_df, pIR, pAQ)
}
\arguments{
\item{input_list}{list of data frames split by fire module}
}
\description{
Module based quarantine operations
}
