% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent_fxns.R
\name{mv_agents}
\alias{mv_agents}
\title{Move agents between fires}
\usage{
mv_agents(agent_df, inc_req, mod_req, t, eir = NULL)
}
\arguments{
\item{agent_df}{data frame containing agent data}

\item{inc_req}{data frame containing incident requests for agents}

\item{mod_req}{data frame containing module requests for agents}

\item{t}{numeric time value}

\item{eir}{Entry Infection Rate, numeric 0<=x<=1, or NULL}
}
\description{
Move agents between fires
}
