% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incident_sampling.R
\name{sample_inc_id}
\alias{sample_inc_id}
\title{Sample specific incident data from seasonal data}
\usage{
sample_inc_id(inc_id_data, inc_info_data, inc_name)
}
\arguments{
\item{inc_id_data}{seasonal incident id assignment data}

\item{inc_info_data}{seasonal incident information data}

\item{inc_name}{desired incident to extract, must be the same case as
\code{inc_info_data$inc_name} (so likely in all caps)}
}
\value{
data frame; only columns and rows that contain the desired incident
}
\description{
Sample specific incident data from seasonal data
}
\examples{
\dontrun{
inc_id_highline <- sample_inc_id(inc_id_2017, inc_info_2017, 'HIGHLINE')
}
}
