% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fxns.R
\name{single_plot}
\alias{single_plot}
\title{Plot runs of a single season}
\usage{
single_plot(data, ...)
}
\arguments{
\item{data}{Outputs from the simulation function}

\item{...}{parameters to pass on to ggplot labs}
}
\description{
Plot runs of a single season
}
