% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_fxns.R
\name{update_eir}
\alias{update_eir}
\title{Update the default EIR list}
\usage{
update_eir(input_df, custom_rates, eir)
}
\arguments{
\item{custom_rates}{Named list containing new infection rates to be applied}

\item{eir_ls}{named list containing infection rates (list(gacc = rate))}
}
\description{
Update the default EIR list
}
