% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent_fxns.R
\name{agents_init}
\alias{agents_init}
\title{Initialize agent data frame for simulation input}
\usage{
agents_init(
  inc_assignments,
  mod_assignments,
  inc_info,
  time = 1,
  leads,
  p_overhead_leads,
  R_init = 0,
  I_init = 0,
  vax_init = 0,
  vax_efficacy = 0.95
)
}
\arguments{
\item{inc_assignments}{Daily incident assignment data}

\item{mod_assignments}{Daily module assignment data}

\item{inc_info}{Incident information for a whole season}

\item{time}{Which day of the season should the agents start on}

\item{leads}{Number of leads each module should have}

\item{p_overhead_leads}{Proportion of overhead modules that should be leaders}

\item{R_init}{Proportion of the population that has already recovered}

\item{I_init}{Proportion of the population that enters the season infected}

\item{vax_init}{Proportion of the population that enters the season vaccinated}

\item{vax_efficacy}{What proportion of those vaccinated actually become immune}
}
\value{
Initialized agent data frame ready for simulation
}
\description{
Initialize agent data frame for simulation input
}
