% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent_fxns.R
\name{assign_roles}
\alias{assign_roles}
\title{Assign roles to agents}
\usage{
assign_roles(input_df, max_leads)
}
\arguments{
\item{input_df}{data frame containing agent data}

\item{max_leads}{number of leads per module}
}
\description{
Assign roles to agents
}
