% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent_fxns.R
\name{assign_roles}
\alias{assign_roles}
\title{Assign roles to agents}
\usage{
assign_roles(df_in, n)
}
\arguments{
\item{df_in}{data frame containing agent data}

\item{n}{number of leads per module}
}
\description{
Assign roles to agents
}
