% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent_fxns.R
\name{assign_roles}
\alias{assign_roles}
\title{Assign roles to agents}
\usage{
assign_roles(input_df, leads, p_overhead_leads)
}
\arguments{
\item{input_df}{Data frame containing agent data}

\item{leads}{Number of leads per module}

\item{p_overhead_leads}{Proportion of overhead to assign as leads}
}
\value{
Vector of agents that need to be assigned as leaders
}
\description{
Assign roles to agents
}
