% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean_up.R
\name{clean_gacc}
\alias{clean_gacc}
\title{Clean up missing gaccs}
\usage{
clean_gacc(inc_data, inc_info)
}
\arguments{
\item{inc_data}{data frame containing daily incident assignments}

\item{inc_info}{data frame containing information on all the incidents of a season}
}
\description{
Agents with missing home gaccs have their home gacc assigned to the gacc of
the first fire of the season.
}
