% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_gaccs.R
\name{clean_gacc}
\alias{clean_gacc}
\title{Clean up res_gaccs == "None"}
\usage{
clean_gacc(data, inc_info)
}
\arguments{
\item{data}{Daily res data frame to clean up}
}
\description{
Clean up res_gaccs == "None"
}
