% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_gaccs.R
\name{clean_gacc}
\alias{clean_gacc}
\title{Clean up res_gaccs == "None"}
\usage{
clean_gacc(inc_data, inc_info)
}
\arguments{
\item{inc_data}{data frame containing daily incident assignments}

\item{inc_info}{Data frame containing information on all the incidents in a season}
}
\description{
Clean up res_gaccs == "None"
}
