% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean_up.R
\name{clean_mods}
\alias{clean_mods}
\title{Group all overhead modules for each incident into 1 module}
\usage{
clean_mods(mod_id_df)
}
\arguments{
\item{mod_id_df}{module id assignment data frame}
}
\value{
all overhead assignments are changed to "O-100"
}
\description{
Solves the problem where incidents had too many unique overhead
modules by changing the all overhead modules to "O-100".
}
