% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent_fxns.R
\name{clean_mods}
\alias{clean_mods}
\title{Group overhead modules into 1 module}
\usage{
clean_mods(mod_ids)
}
\arguments{
\item{mod_ids}{character vector of mod_ids}
}
\description{
Group overhead modules into 1 module
}
\details{
Solves the problem where incidents had too many unique overhead
modules by changing the all overhead modules to "O-100".
}
