% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent_fxns.R
\name{clean_mods}
\alias{clean_mods}
\title{Group overhead modules into 1 module}
\usage{
clean_mods(agent_df)
}
\arguments{
\item{agent_df}{data frame containing firefighter agents data}
}
\description{
Group overhead modules into 1 module
}
