% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean_up.R
\name{count_daily_inf}
\alias{count_daily_inf}
\title{Count daily new infections}
\usage{
count_daily_inf(new_df, agent_df)
}
\arguments{
\item{new_df}{data frame from current/new day}

\item{agent_df}{data frame from previous day}
}
\description{
Count daily new infections
}
