% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_fxns.R
\name{expose_leads}
\alias{expose_leads}
\title{Exposes agents based on assigned roles}
\usage{
expose_leads(inc_df, params)
}
\arguments{
\item{inc_df}{data frame containing agent data for a single inc}

\item{params}{list of parameters for infection dynamics}
}
\description{
Exposes agents based on assigned roles
}
