% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_fxns.R
\name{expose_leads}
\alias{expose_leads}
\title{Find exposed leaders by unique fire (inc_id)}
\usage{
expose_leads(input_df, ...)
}
\arguments{
\item{input_df}{Input data frame}

\item{...}{Exposure params to pass to find_exposed_in_leads}
}
\value{
Vector of resource ID's that have been newly exposed
}
\description{
Find exposed leaders by unique fire (inc_id)
}
