% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_fxns.R
\name{expose_leads}
\alias{expose_leads}
\title{Exposes agents based on assigned roles}
\usage{
expose_leads(inc_df, new_df, BI, BA, exp_thres, delta_t)
}
\arguments{
\item{inc_df}{Single incident data frame. The template for calculating the new data frame.}

\item{new_df}{The data frame to be modified}

\item{BI}{Beta I}

\item{BA}{Beta A}

\item{exp_thres}{How many exposures before infection, default is 1.}

\item{delta_t}{Time step}
}
\description{
Exposes agents based on assigned roles
}
