% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_fxns.R
\name{expose_modules}
\alias{expose_modules}
\title{Find exposed agents by unique module (inc_id x mod_id)}
\usage{
expose_modules(input_df, ...)
}
\arguments{
\item{input_df}{Input data frame}

\item{...}{Exposure params to pass to find_exposed_in_mod; BA, BI, exp_thres, delta_t}
}
\value{
Vector of resource ID's that have been newly exposed
}
\description{
Find exposed agents by unique module (inc_id x mod_id)
}
