% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_fxns.R
\name{expose_modules}
\alias{expose_modules}
\title{Find exposed agents by unique module (inc_id x mod_id)}
\usage{
expose_modules(input_df, BA, BI, exp_thres = 0, delta_t = 1)
}
\arguments{
\item{input_df}{Input data frame}

\item{BA}{Beta for Asymptomatic}

\item{BI}{Beta for Symptomatic}

\item{exp_thres}{Exposure threshold}

\item{delta_t}{Time step}
}
\value{
Vector of resource ID's that have been newly exposed
}
\description{
Find exposed agents by unique module (inc_id x mod_id)
}
