% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_fxns.R
\name{expose_modules}
\alias{expose_modules}
\title{Exposes agents based on assigned modules}
\usage{
expose_modules(inc_df, new_df, BA, BI, exp_thres, delta_t)
}
\arguments{
\item{inc_df}{Single incident data frame. The template for calculating the new data frame.}

\item{new_df}{The data frame to be modified}

\item{BA}{Beta A}

\item{BI}{Beta I}

\item{exp_thres}{How many exposures before infection, default is 1.}

\item{delta_t}{Time step}
}
\description{
Exposes agents based on assigned modules
}
