% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_fxns.R
\name{expose_off_fire}
\alias{expose_off_fire}
\title{Find agents that were exposed off-fire}
\usage{
expose_off_fire(input_df, eir)
}
\arguments{
\item{input_df}{Complete agent data frame (all incidents and modules)}

\item{eir}{Entry infection rate, aka off fire exposure rate}
}
\value{
Vector of resource ID's that have been exposed off-fire
}
\description{
Find agents that were exposed off-fire
}
