% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_fxns.R
\name{find_exposed_in_leads}
\alias{find_exposed_in_leads}
\title{Find exposed leads within a single incident}
\usage{
find_exposed_in_leads(inc_df, BI, BA, exp_thres = 0, delta_t = 1)
}
\arguments{
\item{inc_df}{data frame containing agent information for one incident}

\item{BI}{Beta for Symptomatic}

\item{BA}{Beta for Asymptomatic}

\item{exp_thres}{Exposure threshold}

\item{delta_t}{Time step}
}
\value{
vector of resource ID's to be exposed
}
\description{
Find exposed leads within a single incident
}
