% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_gaccs.R
\name{find_first_mob}
\alias{find_first_mob}
\title{Find the first mobilization for an agent/res_id}
\usage{
find_first_mob(res_data)
}
\arguments{
\item{res_data}{A data frame with}
}
\description{
Find the first mobilization for an agent/res_id
}
