% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modular_quarantine.R
\name{modular_quarantine}
\alias{modular_quarantine}
\title{Module based quarantine operations}
\usage{
modular_quarantine(inc_df, new_df)
}
\arguments{
\item{inc_df}{data frame containing the data for a single incident}

\item{new_df}{data frame to be modified}
}
\description{
Module based quarantine operations
}
