% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modular_quarantine.R
\name{modular_quarantine}
\alias{modular_quarantine}
\title{Module based quarantine operations}
\usage{
modular_quarantine(input_df, pIQ, pAQ)
}
\arguments{
\item{input_df}{the agent data frame at time t}

\item{pIQ}{proportion of infectious symptomatic agents that do not recognize
symptoms and therefore do not quarantine}

\item{pAQ}{proportion if infections asymptomatic agents that are caught by
testing or camp screening and therefore quarantine}
}
\value{
list of res_ids that will be quarantined at time t+1
}
\description{
Module based quarantine operations
}
