% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccine_fxns.R
\name{plan_vaccines}
\alias{plan_vaccines}
\title{Plan vaccines based on gacc information}
\usage{
plan_vaccines(table, x, y)
}
\arguments{
\item{table}{contigency table created by "table" containing frequency of
either gaccs or roles}

\item{x}{numeric vector, time steps}

\item{y}{numeric vector, values from 0-1 specifying proportion of each gacc
to vaccinate at times corresponding to x}
}
\value{
Vaccine plan where row index = time, colnames = gacc, values = number
of agents to vaccinate
}
\description{
Uses the approx function to calculate the number of agents that need to be
vaccinated during each time step rounded to the nearest whole number.
}
