% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_model.R
\name{seasonal_sim}
\alias{seasonal_sim}
\title{Seasonal agent-based model simulation}
\usage{
seasonal_sim(
  inc_data,
  mod_data,
  inc_info,
  incl_A = TRUE,
  De = 5,
  gamma = 8,
  eir = 0.005,
  custom_eir = "none",
  R_init = 0,
  I_init = 0,
  vax_init = 0,
  vax_rate = 0,
  varying_vax = NULL,
  vax_efficacy = 0.95,
  vax_df = NULL,
  R0 = 1.4,
  max_leads = 2,
  prop_overhead_leads = 0.5,
  pIR = 0.2,
  pAQ = 0
)
}
\arguments{
\item{inc_data}{Resource assignments for incident IDs}

\item{mod_data}{Resource assignments for module IDs}

\item{inc_info}{Additional information on each incident}

\item{incl_A}{Include Asymptomatic, TRUE or FALSE}

\item{De}{Incubation period}

\item{gamma}{Recovery time in days}

\item{eir}{Entry infection rate}

\item{custom_eir}{Named list of customized GACC infection rates}

\item{R_init}{Initial number of recovered agents}

\item{I_init}{Initial number of infectious agents}

\item{vax_init}{Initial number of vaccinated agents}

\item{vax_rate}{Percent of total population to vaccinate at each time step}

\item{varying_vax}{Named list specifying a gacc and its new vax rate, for example list(gacc = "NM-SWC", rate = 0.2)}

\item{vax_efficacy}{Proportion of vaccinated agents that become immune}

\item{vax_df}{Data frame with specified number of agents to vaccinate at each time step}

\item{R0}{Basic reproduction parameter}

\item{max_leads}{Number of desired leads per module}

\item{prop_overhead_leads}{Proportion of the overhead module that should be leaders}

\item{pIR}{Probability that Symptomatic agents do not quarantine}

\item{pAQ}{Probability that Asymptomatic will be caught and quarantined}
}
\description{
Seasonal agent-based model simulation
}
