% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_model.R
\name{seasonal_sim}
\alias{seasonal_sim}
\title{Seasonal agent-based model simulation}
\usage{
seasonal_sim(
  inc_data,
  mod_data,
  inc_info,
  incl_A = TRUE,
  De = 5,
  gamma = 10,
  eir = 0.005,
  R_init = 0,
  R0 = 1.4,
  pIR = 0.2,
  pAQ = 0.3,
  I_init = NULL
)
}
\arguments{
\item{inc_data}{Resource assignments for incident IDs}

\item{mod_data}{Resource assignments for module IDs}

\item{inc_info}{Additional information on each incident}

\item{incl_A}{Include Asymptomatic, TRUE or FALSE}

\item{De}{Incubation period}

\item{gamma}{Recovery time in days}

\item{eir}{Entry infection rate}

\item{R_init}{Initial number of recovered agents}

\item{R0}{Basic reproduction parameter}

\item{pIR}{Probability that Symptomatic agents do not quarantine}

\item{pAQ}{Probability that Asymptomatic will be caught and quarantined}

\item{I_init}{Parameters for setting initial infective agents}
}
\description{
Seasonal agent-based model simulation
}
