% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccine_fxns.R
\name{vaccinate}
\alias{vaccinate}
\title{Find out which agents to vaccinate using gacc information}
\usage{
vaccinate(input_df, overhead_ids = NULL, method, plan, efficacy)
}
\arguments{
\item{input_df}{Data frame containing agent roster}

\item{overhead_ids}{Res ids of overhead agents for the purpose of vaccination}

\item{method}{Method of vaccine delivery, either by role or by gacc}

\item{plan}{Vaccination plan}

\item{efficacy}{Proportion of vaccinated agents the gain immunity}
}
\value{
Nested list of res_ids: output$vaccinated and output$immune
}
\description{
Find out which agents to vaccinate using gacc information
}
